raised compiler warning level and fixes all warnings; code refactoring and tests

+++ src/isyntax-io.c
@@ -104,7 +104,7 @@ int32_t anonymize_isyntax_metadata(file_t *fp, int32_t header_size) {
 }
 
 // remove label image and macro image
-int32_t wipe_isyntax_image_data(file_t *fp, int32_t header_size, char *image_type) {
+int32_t wipe_isyntax_image_data(file_t *fp, size_t header_size, char *image_type) {
 
     // gets only the xml header
     char *buffer = (char *)malloc(header_size);
@@ -206,7 +206,7 @@ int32_t handle_isyntax(const char **filename, const char *new_label_name, bool k
         return -1;
     }
 
-    int32_t header_size = get_size_to_substring(fp, ISYNTAX_EOT);
+    size_t header_size = get_size_to_substring(fp, ISYNTAX_EOT);
 
     // remove label image
     int32_t result = wipe_isyntax_image_data(fp, header_size, PHILIPS_LABELIMAGE);
